New prefs system
In case my previous message didn't go through somehow...I added a new, flexible, and fairly slick Preferences system. What I don't yet know is how much overhead it adds. It might be bloating the code of the PRG file.
If so, it will all have to come out. Here it is in the meantime, rhough.